+++ src/app/[locale]/(headerLayout)/like/layout.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+export default function Layout({
+  children,
+  modal,
+}: {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+      {modal}
+    </>
+  );
+}
